+++ include/sys/Utils.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <string>
+#include <vector>
+
+namespace utils {
+    std::vector<std::string> get_env_values(const std::string& name);
+}